+++ index.html
@@ -232,7 +232,10 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 <!--Добавили скрипт-->
+  <script src="main.js" type="module"></script>
+  <script src="data.js" type="module"></script>
+  <script src="util.js" type="module"></script>
   <script src="functions.js"></script>
-  <script src="main.js"></script>
+  
 </body>
 </html>
